+++ plugins/TimeFlow/timeflow.js
@@ -94,10 +94,10 @@ loop();
 function timeflowInfo(){
 	 createDialog({
     id:"timeflow_info",
-    title:selected_language['About']+" Timeflow" ,
+    title:current_config.language['About']+" Timeflow" ,
     content:"TimeFlow collets the time you spent on coding by saving <strong>locally</strong> the data.",
     buttons:{
-      [selected_language['Accept']]:"closeDialog(this)",
+      [current_config.language['Accept']]:"closeDialog(this)",
     }
 
   })