+++ CHANGELOG.md
@@ -1,6 +1,15 @@
 #Graviton Changelog
 This file contains "important" commits, small ones will not be added.
 
+###190413 - 1 [0.7.3]
+- Unified all the configurations in one object
+- Cleaned up code
+- Added a new message on the editors part
+- New Dark theme highlighting (forked from Michael Kaminsky)
+- Enabled MarkDown for codemirror
+- Small bug fixes
+
+
 ###190412 - 1 [0.7.3]
 - Fixed "close" (X) button is hidding when the zoom was the smallest one.
 - Cleaned up code